Try to prevent invalid Alphabet generic params
The Alphabet parameter to Base16 should not contain any repeated
characters, and if Case_Sensitive is not true this test should be case
insensitive. A compile-time error pragma has been added that uses
quantified expressions to test this. Unfortunately, while it uses a
static expression as specified in the UG, GNAT GPL 2015 seems to be
ignoring it silently. It works if modified to be a run-time test...

It has been left in, in the hope that future versions of GNAT may
actually test this properly.