Add support for RFC4648 homoglyphs in Base32
The RFC suggests that for Base32, some users may require '0' to be read
as 'o' or 'O' and '1' to be read as 'l' or 'I' (depending on the case).
When Base32 data is being entered into a computer from a print-out,
these sets of characters may be easily confused due to their visual
similarity.