Add ASCII85 pkg with support for Z85
The ZeroMQ Z85 version of ASCII85 is the most sane of the multiple
variantes, so this is the first variant supported. There is a slight
extension in that if the data is not a multiple of 4 bytes additional
zero bytes are added but then the character output is shortened by the
number of bytes added - so a single byte 'left over' when the encoding
is complete will become two characters output.