Add Empty prim op and tighten post-conditions
The Empty function indicates if there is any data stored in a Codec from
a previous Process operation that could not be processed at that time
because it did not represent a full input group. This allows more
specific post-conditions to be written, which helps in proving whether
output arrays are sufficiently large.