Suppress useless warnings in Base16.ads
These relate to Codec parameters which are not used, or not read. In
some cases this is because the parameter is added only to allow dynamic
dispatch. In others because Base16 To_String conversions do not
require any state to be maintained, but the LSP requires that the
parameter suggest that there is state being read in order to conform to
the other Codec types.